d/control: Re-add versioned dependency on dpkg-dev to make lintian happy.
$ lintian-info --tag build-depends-on-build-essential-package-without-using-version
E: build-depends-on-build-essential-package-without-using-version
N:
N:   The package declares a build-depends on a build essential package
N:   without using a versioned depends. Packages do not have to depend on
N:   any package included in build-essential. It is the responsibility of
N:   anyone building packages to have all build-essential packages
N:   installed. The only reason for an explicit dependency on a package
N:   included in build-essential is if a particular version of that package
N:   is required, in which case the dependency should include the version.
N:
N:   Refer to Debian Policy Manual section 4.2 (Package relationships) for
N:   details.
N:
N:   Severity: important, Certainty: certain
N:
N:   Check: fields, Type: binary, udeb, source