encrypted-boot.md: Replace `luksChangeKey` with `luksConvertKey` for PBKDF downgrade.
`luksConvertKey` is only valid for LUKS2 (and appeared in 2.0.3) but
that's precisely the use-case here.

`luksConvertKey` is nicer than `luksChangeKey` as it preserves the
passphrase and only change the PBKDF algorithm and/or parameters.

Also, drop `--key-slot 0` as we don't care about the index used by the
updated key-slot here.  Moreover — as written in the manpage — in-place
overwriting is unsafe and might render the LUKS header unreadable.